different updates
- merge file-collector and sakura-parsing class into one
- changed error-messages to tables with the table-items
