+++ tests/test_strings/branch_test_string.h
@@ -32,7 +32,7 @@ std::string testBranchString = "[install_branch]\n"
                                "{\n"
                                "    apt (\"this is a test\")\n"
                                "    -> update \n"
-                               "    -> present:\n"
+                               "    -> pres ent:\n"
                                "       - names = \"{{packages}}\"\n"
                                "\n"
                                "    for(i = 0; i < packages.size(); i++)"
